+++ Fusion/Algorithms/Ecofer/Data/Model/DynamicOutput.cs
@@ -7,30 +7,30 @@ namespace Data.Model
 {
     public class DynamicOutput
     {
-        public float m_Tavby;
-        public float E_Tavby;
-        public float T_Tavby;
-        public float FP_C;
-        public float[] FP_Tavby;
-        public float[] m_SlozkaTavby;
+        public double m_Tavby;
+        public double E_Tavby;
+        public double T_Tavby;
+        public double FP_C;
+        public double[] FP_Tavby;
+        public double[] m_SlozkaTavby;
 
-        public float m_Struska;
-        public float m_Kov;
-        public float[] FP_Struska;
-        public float[] FP_Kov;
-        public float[] m_SlozkaStruska;
-        public float[] m_SlozkaKov;
-        public float[] c_Struska;
-        public float[] c_Kov;
+        public double m_Struska;
+        public double m_Kov;
+        public double[] FP_Struska;
+        public double[] FP_Kov;
+        public double[] m_SlozkaStruska;
+        public double[] m_SlozkaKov;
+        public double[] c_Struska;
+        public double[] c_Kov;
 
         public DateTime StartTime;
         public TimeSpan Duration;
         public DateTime ActualTime;
 
-        public float E_C_oxidace;
-        public float E_Si_oxidace;
-        public float E_Mn_oxidace;
-        public float E_Al_oxidace;
-        public float E_Fe_oxidace;
+        public double E_C_oxidace;
+        public double E_Si_oxidace;
+        public double E_Mn_oxidace;
+        public double E_Al_oxidace;
+        public double E_Fe_oxidace;
     }
 }
